safari-534.53-branch/LayoutTests/http/tests/misc/frame-default-enc-different-domain-expected.txt

ALERT: ISO-8859-1
